Add files via upload
The summary of the Dashboard presentation are as follows:

The number of ATM transactions has consistently increased over the years, making it the primary mode of monetary transactions in Nigeria for 11 consecutive years.

The number of cheque transactions has significantly decreased over the years, yet it holds the highest cumulative transaction amount, accounting for 67.18% of the total transactions made in Nigeria from January 2009 to December 2019.


At 438,614,182, 2019 had the highest Number of POS transactions and was 47,666.0% higher than 2009, which had the lowest Number of POS transactions at 918,256.

2019 accounted for 43.21% of Number of POS transactions.

Across all 11 years, Number of Cheque transactions ranged from 7,275,237 to 37,718,585, Number of POS transactions ranged from 918,256 to 438,614,182

Conclusion
Electronic transaction methods like ATMs and POS have garnered significant popularity among Nigerians. The challenges of waiting in queues at banks and the limited staffing in many branch locations have contributed to the increasing adoption of electronic payment methods over time. In an era where a substantial portion of the economy is becoming digitalized, Nigerians should strive to become more acquainted with electronic payment methods, at the very least.

Thank you for reading.